% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot the path on a leaflet map}
\usage{
plot_map(d, colvar = "T_Depth")
}
\arguments{
\item{d}{Data.frame containing all data (must have Lat/Lon/T_Depth columns)}

\item{colvar}{Variable to color path points (defaults to T_Depth)}
}
\description{

}
\examples{

data(greenlake)

plot_map(greenlake)

#the map can be colored by different variables
plot_map(greenlake, colvar='pH')


}

