% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot heat map of overall profile}
\usage{
plot_profile(d, colvar = "Sled_YSI_Temp", xvar = "RECORD",
  yvar = "Sled_Depth", col = rev(heat.colors(24)), fig.out = NULL,
  tau = 0)
}
\arguments{
\item{d}{Data.frame containing all data (must have Lat/Lon/T_Depth columns)}

\item{colvar}{Variable to color path points (defaults to T_Depth)}

\item{xvar}{Select a different variable to define the x-axis}

\item{yvar}{Select a different variable to define the y-axis}

\item{col}{Select different colormap (passed in as vector of colors, for example, try \code{rainbow(12)})}

\item{tau}{Rudimentary sensor lag correction. Zero (no correction) is default. Larger value applies 
larger lag correction, though it adds noise at the same time}
}
\description{
Creates single heat map of a chosen variable
}
\examples{
data(greenlake)

plot_profile(greenlake)


#Can change the heatmap variable if you want
plot_profile(greenlake, colvar='Sled_chl_A')


}

